+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 * Feat: Add option to ignore exceptions by type (#1352)
+* Feat: Sentry closes Android NDK and ShutdownHook integrations
 
 # 4.4.0-alpha.1